+++ CHANGELOG.md
@@ -7,7 +7,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
-None.
+### Changed
+
+- Remove configuration option `onSave` as this is the default and probably not toggled by anyone
 
 ## [0.3.0] - 2021-03-27